Move coverage phase to before FirstTransform and handle more trees
